contains duplicate

+++ arrays/besttimetobuysellstocks/OptimalApproach.py
@@ -4,7 +4,7 @@ from typing import List
 
 def get_max_profit(prices: List[int]) -> int:
     max_profit = 0
-    min_price = int('inf')
+    min_price = float('inf')
 
     for price in prices:
         min_price = min(price, min_price)
